Moves the components to the src directory
