+++ packages/components/index.js
@@ -11,38 +11,38 @@ export { default as Button } from "./Button";
 export { default as ErrorBoundary } from "./ErrorBoundary";
 export { default as Explanation } from "./Explanation";
 export { default as Heading } from "./Heading";
-export { HelpCenterButton } from "./HelpCenterButton";
-export { default as HelpText } from "./HelpText";
-export { default as Icon } from "./Icon";
-export { default as IconButtonToggle } from "./IconButtonToggle.js";
+export { HelpCenterButton } from "./src/HelpCenterButton";
+export { default as HelpText } from "./src/HelpText";
+export { default as Icon } from "./src/Icon";
+export { default as IconButtonToggle } from "./src/IconButtonToggle.js";
 export { default as IFrame } from "./IFrame";
 export { default as Input } from "./Input";
 export { default as Label } from "./Label";
-export { default as LanguageNotice, languageNoticePropType } from "./LanguageNotice";
+export { default as LanguageNotice, languageNoticePropType } from "./src/LanguageNotice";
 export { default as Loader } from "./Loader";
 export { default as Notification } from "./Notification";
 export { default as Paper } from "./Paper";
 export { default as Progressbar } from "./Progressbar";
 export { default as Section } from "./Section";
-export { SectionTitle } from "./SectionTitle";
-export { default as ScoreAssessments } from "./ScoreAssessments";
-export { default as StackedProgressBar } from "./StackedProgressBar";
+export { SectionTitle } from "./src/SectionTitle";
+export { default as ScoreAssessments } from "./src/ScoreAssessments";
+export { default as StackedProgressBar } from "./src/StackedProgressBar";
 export { default as SvgIcon } from "./SvgIcon";
 export { default as Textarea } from "./Textarea";
-export { default as Toggle } from "./Toggle";
-export { UpsellButton } from "./UpsellButton";
-export { UpsellLinkButton } from "./UpsellLinkButton";
-export { YoastButton } from "./YoastButton";
-export { YoastInputField } from "./YoastInput";
-export { YoastLinkButton } from "./YoastLinkButton";
+export { default as Toggle } from "./src/Toggle";
+export { UpsellButton } from "./src/UpsellButton";
+export { UpsellLinkButton } from "./src/UpsellLinkButton";
+export { YoastButton } from "./src/YoastButton";
+export { YoastInputField } from "./src/YoastInput";
+export { YoastLinkButton } from "./src/YoastLinkButton";
 export { default as YoastLogo } from "./YoastLogo";
-export { default as YoastModal } from "./YoastModal";
+export { default as YoastModal } from "./src/YoastModal";
 export { default as YoastSeoIcon } from "./YoastSeoIcon";
-export { default as YoastTabs } from "./YoastTabs";
-export { default as YoastWarning } from "./YoastWarning";
+export { default as YoastTabs } from "./src/YoastTabs";
+export { default as YoastWarning } from "./src/YoastWarning";
 export { default as YouTubeVideo } from "./YouTubeVideo";
 
-export { default as SynonymsInput } from "./composites/SynonymsInput";
+export { default as SynonymsInput } from "./src/SynonymsInput";
 export { default as Textfield } from "./composites/Textfield";
 
 export { getRtlStyle } from "./utils/helpers/styled-components";
