+++ packages/components/src/StackedProgressBar.js
