+++ packages/components/src/YoastInput.js
@@ -2,7 +2,7 @@
 import styled from "styled-components";
 
 // Internal dependencies.
-import { colors, rgba } from "./style-guide";
+import { colors, rgba } from "../style-guide";
 
 export const YoastInputContainer = styled.div`
 	display: flex;
