+++ packages/components/tests/HelpTextTest.js
@@ -3,7 +3,7 @@ import React from "react";
 import renderer from "react-test-renderer";
 
 /* Internal dependencies */
-import HelpText from "../HelpText";
+import HelpText from "../src/HelpText";
 
 describe( "HelpText", () => {
 	it( "matches the snapshot by default", () => {
